quicksave

+++ insert.php
@@ -1,18 +1,26 @@
 <?php
 require_once 'functions.php';
-# bulk insert process per file, to be run in parallel
+# bulk insert process per file, to be run in parallel.
+# this script checks for validity of all comment rows in an archive, manually builds the query,
+# then batch insert them to the database using prepared statements.
 
 $file = $argv[1];
-$tags = get('tags');
+$tags = val('tags');
 
-$db = getConnection_db(get('db_name'));
+@$db = getConnection_db(val('db_name')) or die ("Couldn't connect: ".$db->errno);
 if (!(@$fp = fopen($file, 'rb'))) {
-    cleanup_process($db);
-    exit ("File $file could not be opened");
+    cleanup_process($db, $file);
+    exit ("File $file could not be opened\n");
 }
 
-$tags = get('tags');
-$buffer_size = get('insert_batch');
+create_compressed_table($db);
+
+# a couple of tweaks to make the insert process faster
+# http://derwiki.tumblr.com/post/24490758395/loading-half-a-billion-rows-into-mysql
+optimize($db);
+
+$tags = val('tags');
+$buffer_size = val('insert_batch');
 $rows = 0;
 $start = microtime(true);
 const value_types = ['BOOL', 'VARCHAR', 'PRIMARY KEY'];
@@ -43,9 +51,7 @@ $lookup = array_keys($tags);
 # note that the data is ridiculously 'dirty', so we'll be doing a lot of manual checks!
 # used to check for any duplicate primary keys
 $primary_keys = [];
-
-# stage transaction, since we disabled auto-commit
-//$db->query('START TRANSACTION');
+$num_queries = 0;
 
 while (!feof($fp)) {
     # read buffer-size lines at a time - the first parameter will be the type string
@@ -53,25 +59,19 @@ while (!feof($fp)) {
         @$comment = json_decode(fgets($fp), true);
         # because the data is so fucked, sometimes a line of file isn't even a proper json
         if (!is_array($comment)) continue;
-        # build query
-        $prep_stmt = $prep_stmt.$question_marks;
-        # extend type string
-        $params[0] = $params[0].$type_string;
         # prepare values to bind for each comment
         foreach ($comment as $tag => $value) {
             # see if this is a tag we care about
             if (($key = array_search($tag, $lookup)) === false) continue;
             $valid = true;
             # check the validity of each entry, according to its tag
-            if (array_key_exists($tag, $type_restraint)) {
+            if (array_key_exists($tag, $type_restraint))
                 switch ($type_restraint[$tag]) {
                     case 'PRIMARY KEY':
                         # search for any duplicate primary key
-                        if (array_search($value, $primary_keys) !== false) {
-                            $valid = false;
-                        } else {
-                            $primary_keys[] = $value;
-                        }
+                        array_search($value, $primary_keys) !== false
+                                ? $valid = false
+                                : $primary_keys[] = $value;
                         break;
                     case 'VARCHAR':
                         # the input strings shouldn't be longer than the varchar length limit
@@ -83,15 +83,13 @@ while (!feof($fp)) {
                         if (array_search($value, valid_bool) === false) $valid = false;
                         break;
                 }
-            }
             if ($valid) {
+//                if (!val('prep_stmt') && is_string($value))
+//                    $value = mysqli_real_escape_string($db, $value);
                 # we put all the values into one array, so we need to take care of the relative order
                 $params[$num_fields * $i + $key + 1] = $value;
                 continue;
             }
-            # invalidate this entire comment row if there's an error
-            $prep_stmt = substr($prep_stmt, 0, -strlen($question_marks));
-            $params[0] = substr($params[0], $num_fields);
             # delete the values for this comment row
             for ($j = $num_fields * $i + 1; $j <= ($i + 1) * $num_fields; $j++)
                 if (isset($params[$j])) unset($params[$j]);
@@ -100,29 +98,56 @@ while (!feof($fp)) {
             break;
         }
     }
+    
+    for ($k = 0; $k < $i; $k++) {
+        if (val('prep_stmt')) {
+            # build query
+            $prep_stmt = $prep_stmt.$question_marks;
+            # extend type string
+            $params[0] = $params[0].$type_string;
+        } else {
+            $prep_stmt = $prep_stmt.'(';
+            # $num_fields * $i + $key + 1
+            for ($h = 0; $h < $num_fields; $h++)
+                $prep_stmt = $prep_stmt.$params[$num_fields * $k + $h + 1].',';
+            $prep_stmt = rtrim($prep_stmt, ',').'),';
+        }
+    }
+    $prep_stmt = rtrim($prep_stmt, ',');
     $rows += $i;
     
     try {
-        # execute the giant insert query
-        $prep_stmt = rtrim($prep_stmt, ',');
-        @$stmt = $db->prepare($prep_stmt) or die ('Failed to prepare: '.$db->errno);
-        # to bind params in order, we need to sort them, then hand in the reference via refValues
-        ksort($params);
-        @call_user_func_array([$stmt, 'bind_param'], refValues($params));
-        $stmt->execute();
-    } catch (Exception $e) {
-        cleanup_process($db);
-        exit("Failed to execute query. Error $e");
+        if (val('prep_stmt')) {
+            # execute the giant insert query
+            @$stmt = $db->prepare($prep_stmt) or die ('Failed to prepare: '.$db->errno."\n");
+            # to bind params in order, we need to sort them, then hand in the reference via refValues
+            ksort($params);
+            @call_user_func_array([$stmt, 'bind_param'], refValues($params));
+            $stmt->execute();
+        } else {
+            # stage transaction, since we disabled auto-commit
+            if (!$num_queries++)
+                $db->query('START TRANSACTION');
+            $db->query($prep_stmt);
+            if ($num_queries == val('num_trans') || feof($fp)) {
+                $db->query('COMMIT');
+                if ($num_queries == val('num_trans')) $num_queries -= val('num_trans');
+            }
+        }
+    } catch (Error $e) {
+        cleanup_process($db, $file);
+        exit("Failed to execute query. Error $e\n");
     }
+    echo "rows: $rows\n";
 }
 
-//$db->query('COMMIT');
-exec("rm $file");
+if (val('cleanup')) exec("rm $file");
 
 $num_rows = $db->query("SELECT COUNT(*) FROM Comments")->fetch_row()[0];
-$duration = (int) microtime(true) - $start;
-cleanup_process($db);
-exit ("Inserted $num_rows rows out of $rows from $file in $duration s.");
+cleanup_process($db, $file);
+$duration = (int) (microtime(true) - $start);
+printf('Inserted %d out of %d rows from [%s] in [%.6f]s.\n', $num_rows, $rows, $file, $duration);
+exit();
 
 # http://php.net/manual/en/mysqli-stmt.bind-param.php#100879
 function refValues($arr){
